+++ R/esp_get_capimun.R
@@ -143,7 +143,7 @@ esp_get_capimun <- function(year = Sys.Date(),
 
   if (!is.null(munic)) {
     munic <- paste(munic, collapse = "|")
-    data_sf <- data_sf[grep(munic, data_sf$name), ]
+    data_sf <- data_sf[grep(munic, data_sf$name, ignore.case = TRUE), ]
   }
 
   if (!is.null(region)) {